requirements.txt
blinker==1.4
certifi==2019.11.28
Click==7.0
dnspython==1.16.0
dominate==2.4.0
eventlet==0.28.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.9.1
Flask-MQTT==1.0.5
Flask-SocketIO==4.3.1
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
monotonic==1.5
paho-mqtt==1.5.0
python-engineio==3.13.0
python-socketio==4.6.0
six==1.14.0
typing==3.7.4.1
visitor==0.1.3
Werkzeug==1.0.0
